fixed proxy manager delay issue

+++ src/ScraperFramework.Shared/Pocos/CrawlDescription.cs
@@ -6,11 +6,11 @@
 
         public int KeywordID { get; set; }
 
-        public int SearchEngineID { get; set; }
+        public short SearchEngineID { get; set; }
 
-        public int RegionID { get; set; }
+        public short RegionID { get; set; }
 
-        public int CityID { get; set; }
+        public short CityID { get; set; }
 
         public int SearchEngineGroup { get; set; }
